fix: working sub directory
