+++ .github/workflows/build.yml
@@ -19,10 +19,6 @@ jobs:
         with:
           ref: ${{ github.event.pull_request.head.ref }}
           repository: ${{ github.event.pull_request.head.repo.full_name }}
-      - name: Setup Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: 16.13.1
       - name: Install dependencies
         run: yarn install --check-files
       - name: build
@@ -91,7 +87,7 @@ jobs:
     steps:
       - uses: actions/setup-node@v3
         with:
-          node-version: 16.13.1
+          node-version: 14.x
       - name: Download build artifacts
         uses: actions/download-artifact@v3
         with:
@@ -116,7 +112,7 @@ jobs:
     steps:
       - uses: actions/setup-node@v3
         with:
-          node-version: 16.13.1
+          node-version: 14.x
       - uses: actions/setup-python@v4
         with:
           python-version: 3.x
