+++ .github/workflows/release.yml
@@ -24,10 +24,6 @@ jobs:
         run: |-
           git config user.name "github-actions"
           git config user.email "github-actions@github.com"
-      - name: Setup Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: 16.13.1
       - name: Install dependencies
         run: yarn install --check-files --frozen-lockfile
       - name: release
@@ -57,7 +53,7 @@ jobs:
     steps:
       - uses: actions/setup-node@v3
         with:
-          node-version: 16.13.1
+          node-version: 14.x
       - name: Download build artifacts
         uses: actions/download-artifact@v3
         with:
@@ -86,7 +82,7 @@ jobs:
     steps:
       - uses: actions/setup-node@v3
         with:
-          node-version: 16.13.1
+          node-version: 14.x
       - name: Download build artifacts
         uses: actions/download-artifact@v3
         with:
@@ -119,7 +115,7 @@ jobs:
     steps:
       - uses: actions/setup-node@v3
         with:
-          node-version: 16.13.1
+          node-version: 14.x
       - uses: actions/setup-python@v4
         with:
           python-version: 3.x
