+++ .github/workflows/upgrade-main.yml
@@ -18,10 +18,6 @@ jobs:
         uses: actions/checkout@v3
         with:
           ref: main
-      - name: Setup Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: 16.13.1
       - name: Install dependencies
         run: yarn install --check-files --frozen-lockfile
       - name: Upgrade dependencies
