+++ .gitignore
@@ -52,7 +52,7 @@ junit.xml
 tsconfig.json
 !/API.md
 /assets/
-!/src/pstn-function.ts
+!/src/resources/pstn/pstn-function.ts
 cdk.out
 cdk.context.json
 yarn-error.log
