+++ .projen/files.json
@@ -14,7 +14,7 @@
     ".projen/files.json",
     ".projen/tasks.json",
     "LICENSE",
-    "src/pstn-function.ts",
+    "src/resources/pstn/pstn-function.ts",
     "tsconfig.dev.json"
   ],
   "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
