+++ .projen/tasks.json
@@ -46,25 +46,25 @@
       "description": "Prepare assets",
       "steps": [
         {
-          "spawn": "bundle:pstn.lambda"
+          "spawn": "bundle:resources/pstn/pstn.lambda"
         }
       ]
     },
-    "bundle:pstn.lambda": {
-      "name": "bundle:pstn.lambda",
-      "description": "Create a JavaScript bundle from src/pstn.lambda.ts",
+    "bundle:resources/pstn/pstn.lambda": {
+      "name": "bundle:resources/pstn/pstn.lambda",
+      "description": "Create a JavaScript bundle from src/resources/pstn/pstn.lambda.ts",
       "steps": [
         {
-          "exec": "esbuild --bundle src/pstn.lambda.ts --target=\"node18\" --platform=\"node\" --outfile=\"assets/pstn.lambda/index.js\" --tsconfig=\"tsconfig.dev.json\" --external:aws-sdk --sourcemap"
+          "exec": "esbuild --bundle src/resources/pstn/pstn.lambda.ts --target=\"node18\" --platform=\"node\" --outfile=\"assets/resources/pstn/pstn.lambda/index.js\" --tsconfig=\"tsconfig.dev.json\" --external:aws-sdk"
         }
       ]
     },
-    "bundle:pstn.lambda:watch": {
-      "name": "bundle:pstn.lambda:watch",
-      "description": "Continuously update the JavaScript bundle from src/pstn.lambda.ts",
+    "bundle:resources/pstn/pstn.lambda:watch": {
+      "name": "bundle:resources/pstn/pstn.lambda:watch",
+      "description": "Continuously update the JavaScript bundle from src/resources/pstn/pstn.lambda.ts",
       "steps": [
         {
-          "exec": "esbuild --bundle src/pstn.lambda.ts --target=\"node18\" --platform=\"node\" --outfile=\"assets/pstn.lambda/index.js\" --tsconfig=\"tsconfig.dev.json\" --external:aws-sdk --sourcemap --watch"
+          "exec": "esbuild --bundle src/resources/pstn/pstn.lambda.ts --target=\"node18\" --platform=\"node\" --outfile=\"assets/resources/pstn/pstn.lambda/index.js\" --tsconfig=\"tsconfig.dev.json\" --external:aws-sdk --watch"
         }
       ]
     },
