--- layer/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "layer",
-  "version": "1.0.0",
-  "main": "index.js",
-  "license": "MIT",
-  "dependencies": {
-    "@aws-sdk/client-chime-sdk-voice": "^3.226.0",
-    "@aws-sdk/client-ssm": "^3.226.0"
-  }
-}
