+++ src/resources/pstn/phoneNumber.ts
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-extraneous-dependencies */
 import {
   ChimeSDKVoiceClient,
   CreatePhoneNumberOrderCommand,
@@ -27,74 +28,6 @@ const chimeSDKVoiceClient = new ChimeSDKVoiceClient({
 
 const ssmClient = new SSMClient({ region: process.env.AWS_REGION });
 
-import {
-  CdkCustomResourceEvent,
-  CdkCustomResourceResponse,
-  Context,
-} from 'aws-lambda';
-
-const response: CdkCustomResourceResponse = {};
-let resourcePropertiesUid: string;
-let requestProperties: {};
-
-export const handler = async (
-  event: CdkCustomResourceEvent,
-  context: Context,
-): Promise<CdkCustomResourceResponse> => {
-  console.info('event: ', event);
-  const resourceType = event.ResourceProperties.resourceType;
-  const requestType = event.RequestType;
-  resourcePropertiesUid = event.ResourceProperties.uid;
-  requestProperties = event.ResourceProperties.properties;
-
-  response.StackId = event.StackId;
-  response.RequestId = event.RequestId;
-  response.LogicalResourceId = event.LogicalResourceId;
-  response.PhysicalResourceId = context.logGroupName;
-
-  switch (resourceType) {
-    case 'PhoneNumber':
-      switch (requestType) {
-        case 'Create':
-          try {
-            response.Data = await CreatePhoneNumber(
-              resourcePropertiesUid,
-              requestProperties,
-            );
-            response.Status = 'SUCCESS';
-            response.Reason = 'CreatePhoneNumber successful';
-          } catch (error) {
-            if (error instanceof Error) {
-              response.Status = 'FAILED';
-              response.Reason = error.message;
-            }
-          }
-          break;
-        case 'Update':
-          response.Status = 'SUCCESS';
-          break;
-        case 'Delete':
-          try {
-            await DeletePhoneNumber(resourcePropertiesUid);
-            response.Status = 'SUCCESS';
-            response.Reason = 'DeletePhoneNumber successful';
-          } catch (error) {
-            if (error instanceof Error) {
-              response.Status = 'FAILED';
-              response.Reason = error.message;
-            }
-          }
-          break;
-      }
-      break;
-
-    case 'VoiceConnector':
-      break;
-  }
-  console.log(`Response: ${JSON.stringify(response)}`);
-  return response;
-};
-
 let searchAvailableNumbersParam: SearchAvailablePhoneNumbersCommandInput;
 let searchAvailablePhoneNumberResponse: SearchAvailablePhoneNumbersCommandOutput;
 let createPhoneNumberOrderResponse: CreatePhoneNumberOrderCommandOutput;
@@ -111,7 +44,7 @@ export interface CreatePhoneNumberProps {
   phoneProductType?: string;
 }
 
-const CreatePhoneNumber = async (
+export const CreatePhoneNumber = async (
   uid: string,
   props: CreatePhoneNumberProps,
 ) => {
@@ -200,7 +133,7 @@ const CreatePhoneNumber = async (
   };
 };
 
-const DeletePhoneNumber = async (uid: string) => {
+export const DeletePhoneNumber = async (uid: string) => {
   try {
     getParameterCommandOutput = await ssmClient.send(
       new GetParameterCommand({ Name: '/chime/phoneNumber' + uid }),
