+++ src/resources/pstnPython/sipmediaapp.py
@@ -7,7 +7,7 @@ try:
     log_level = os.environ["LogLevel"]
     if log_level not in ["INFO", "DEBUG"]:
         log_level = "INFO"
-except:
+except BaseException:
     log_level = "INFO"
 logger.setLevel(log_level)
 
@@ -73,3 +73,5 @@ def delete_sip_media_app(uid):
         raise RuntimeError(error)
 
     return True
+
+
